bot/requirements.txt
python-telegram-bot[job-queue]==21.9
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
